requirements.txt
celery
cherrypy
cloudevents-python
jsonpath2
pacifica-downloader
pacifica-uploader
peewee
setuptools
six
git+https://github.com/IMMM-SFA/proxymod.git
